Move migrtions into services and base into modules/migration
