+++ modules/migration/messenger.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-package base
+package migration
 
 // Messenger is a formatting function similar to i18n.Tr
 type Messenger func(key string, args ...interface{})
