+++ modules/migration/repo.go
@@ -3,7 +3,7 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-package base
+package migration
 
 // Repository defines a standard repository information
 type Repository struct {
